refactor necessary functions to validators package, added prev epoch helpers
