+++ beacon-chain/core/epoch/epoch_processing.go
@@ -4,10 +4,8 @@ import (
 	"bytes"
 
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/types"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/validators"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	"github.com/prysmaticlabs/prysm/shared/params"
-	"github.com/prysmaticlabs/prysm/shared/slices"
 )
 
 // Attestations returns the pending attestations of slots in the epoch
@@ -15,8 +13,8 @@ import (
 // included in the chain during the epoch.
 //
 // Spec pseudocode definition:
-//   [a for a in state.latest_attestations if state.slot - EPOCH_LENGTH <=
-//   a.data.slot < state.slot]
+//   [a for a in state.latest_attestations
+//   if state.slot - EPOCH_LENGTH <= a.data.slot < state.slot]
 func Attestations(state *pb.BeaconState) []*pb.PendingAttestationRecord {
 	epochLength := params.BeaconConfig().EpochLength
 	var thisEpochAttestations []*pb.PendingAttestationRecord
@@ -39,7 +37,7 @@ func Attestations(state *pb.BeaconState) []*pb.PendingAttestationRecord {
 }
 
 // BoundaryAttestations returns the pending attestations from
-// the epoch boundary block.
+// the epoch's boundary block.
 //
 // Spec pseudocode definition:
 //   [a for a in this_epoch_attestations if a.data.epoch_boundary_root ==
@@ -69,62 +67,78 @@ func BoundaryAttestations(
 	return boundaryAttestations, nil
 }
 
-// BoundaryAttesterIndices returns the attester indices from epoch
-// boundary block.
+// PrevAttestations returns the attestations of the previous epoch
+// (state.slot - 2 * EPOCH_LENGTH...state.slot - EPOCH_LENGTH).
 //
 // Spec pseudocode definition:
-//   Let this_epoch_boundary_attester_indices be the union of the validator
-//   index sets given by [get_attestation_participants(state, a.data, a.participation_bitfield)
-//   for a in this_epoch_boundary_attestations]
-func BoundaryAttesterIndices(
-	state *pb.BeaconState,
-	boundaryAttestations []*pb.PendingAttestationRecord,
-) ([]uint32, error) {
+//   [a for a in state.latest_attestations
+//   if state.slot - 2 * EPOCH_LENGTH <= a.slot < state.slot - EPOCH_LENGTH]
+func PrevAttestations(state *pb.BeaconState) []*pb.PendingAttestationRecord {
+	epochLength := params.BeaconConfig().EpochLength
+	var prevEpochAttestations []*pb.PendingAttestationRecord
+	var earliestSlot uint64
 
-	var attesterIndicesIntersection []uint32
-	for _, boundaryAttestation := range boundaryAttestations {
-		attesterIndices, err := validators.AttestationParticipants(
-			state,
-			boundaryAttestation.Data,
-			boundaryAttestation.ParticipationBitfield)
+	for _, attestation := range state.LatestAttestations {
+
+		// If the state slot is less than 2 * epochLength, then the earliestSlot would
+		// result in a negative number. Therefore we should default to
+		// earliestSlot = 0 in this case.
+		if state.Slot > 2*epochLength {
+			earliestSlot = state.Slot - 2*epochLength
+		}
+
+		if earliestSlot <= attestation.GetData().Slot &&
+			attestation.GetData().Slot < earliestSlot+2*epochLength {
+			prevEpochAttestations = append(prevEpochAttestations, attestation)
+		}
+	}
+	return prevEpochAttestations
+}
+
+// PrevJustifiedAttestations returns the justified attestations
+// of the previous 2 epochs.
+//
+// Spec pseudocode definition:
+//   [a for a in this_epoch_attestations + previous_epoch_attestations
+//   if a.justified_slot == state.previous_justified_slot]
+func PrevJustifiedAttestations(
+	state *pb.BeaconState,
+	thisEpochAttestations []*pb.PendingAttestationRecord,
+	prevEpochAttestations []*pb.PendingAttestationRecord,
+) []*pb.PendingAttestationRecord {
+	var prevJustifiedAttestations []*pb.PendingAttestationRecord
+	epochAttestations := append(thisEpochAttestations, prevEpochAttestations...)
+
+	for _, attestation := range epochAttestations {
+		if attestation.Data.JustifiedSlot == state.PreviousJustifiedSlot {
+			prevJustifiedAttestations = append(prevJustifiedAttestations, attestation)
+		}
+	}
+	return prevEpochAttestations
+}
+
+// PrevHeadAttestations returns the pending attestations from
+// the canonical beacon chain.
+//
+// Spec pseudocode definition:
+//   [a for a in previous_epoch_attestations
+//   if a.beacon_block_root == get_block_root(state, a.slot)]
+func PrevHeadAttestations(
+	state *pb.BeaconState,
+	prevEpochAttestations []*pb.PendingAttestationRecord,
+) ([]*pb.PendingAttestationRecord, error) {
+	var headAttestations []*pb.PendingAttestationRecord
+
+	for _, attestation := range prevEpochAttestations {
+		canonicalBlockRoot, err := types.BlockRoot(state, attestation.Data.Slot)
 		if err != nil {
 			return nil, err
 		}
 
-		attesterIndicesIntersection = slices.Union(attesterIndicesIntersection, attesterIndices)
+		attestationData := attestation.GetData()
+		if bytes.Equal(attestationData.BeaconBlockRootHash32, canonicalBlockRoot) {
+			headAttestations = append(headAttestations, attestation)
+		}
 	}
-
-	return attesterIndicesIntersection, nil
-}
-
-// BoundaryAttesters returns the attesters validator object from epoch
-// boundary block.
-//
-// Spec pseudocode definition:
-//   Let this_epoch_boundary_attesters = [state.validator_registry[i]
-//   for indices in this_epoch_boundary_attester_indices for i in indices].
-func BoundaryAttesters(state *pb.BeaconState, attesterIndices []uint32) []*pb.ValidatorRecord {
-
-	var boundaryAttesters []*pb.ValidatorRecord
-	for _, attesterIndex := range attesterIndices {
-		boundaryAttesters = append(boundaryAttesters, state.ValidatorRegistry[attesterIndex])
-	}
-
-	return boundaryAttesters
-}
-
-// BoundaryAttestingBalance returns the epoch boundary attesting
-// balance.
-//
-// Spec pseudocode definition:
-//   Let this_epoch_boundary_attesting_balance =
-//   sum([get_effective_balance(v) for v in this_epoch_boundary_attesters])
-func BoundaryAttestingBalance(boundaryAttesters []*pb.ValidatorRecord) uint64 {
-
-	var boundaryAttestingBalance uint64
-	for _, boundaryAttester := range boundaryAttesters {
-		boundaryAttestingBalance += validators.EffectiveBalance(boundaryAttester)
-	}
-
-	return boundaryAttestingBalance
+	return headAttestations, nil
 }
